suppression de countListSquareRec

+++ src/common/libmmgcommonf.h
@@ -11,7 +11,7 @@
 ! ** See the "libmmg*.h" file for a correct displaying of the documentation.
 ! */
 
-#define MMG5_DATA_PTR_T INTEGER(kind=8)
+#define MMG5_DATA_PTR_T INTEGER(kind=4)
 ! /* =============================================================================
 ! **  This file is part of the mmg software package for the tetrahedral
 ! **  mesh modification.
